Fix geometry-less exported as empty layers
If the selection by BBOX did not select anything, make sure we fool `QgsOfflineEditing` something is selected.
Otherwise when `layer.selectedFeatureIds().isEmpty()`, `QgsOfflineEditing` dumps all features.
NOTE that `layer.selectedFeatureIds()` should be renamed to `layer.requestedFeaturesIds()`, as it does
not indicate actual features ids being already selected.

This was wrongly applied on non-spatial layers, which caused them to be exported as empty layers.