fix: do not allow "Directly access datasource" on QFieldCloud packaging actions for file based layers that are not readonly
When configuring QFieldCloud packaging options, we might get the confusing scenario of configuring a file based vector layer (gpkg) with "Directly access datasource" which will make the layer editable on QField, but never synchronizable back to QFieldCloud. This is bad and may cause potential data loss.

So we can just disable this option in a way, that "Directly access datasource" is available for :
1) non-file based layers (virtual, postgis, wfs etc)
2) file based layers that are in read-only mode (readonly gpkg)

In this PR I also enabled removing localized and virtual layers from the project.