fix: do not use Enum as default value which causes the plugin to fail after reloading
NOTE: while it will be much more readable to pass `export_type` parameter to have a default value of `ExportType.Cable`,
we cannot reference an Enum as a default value in a QGIS plugin, as when the file reloaded (e.g. enable/disable the plugin, Plugin reloader),
the enum will have a different address and comparisons will fail.